requirements.txt
google-search-results
holidays
python-dotenv
requests
